requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
groq>=0.4.0
langchain-google-genai>=1.0.0
langchain-pinecone>=0.1.0
langchain-community>=0.0.20
pinecone-client>=3.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
openai-whisper
vosk>=0.3.45
pyaudio>=0.2.11
soundfile>=0.12.1
gtts>=2.4.0
spacy>=3.7.0
nest-asyncio>=1.5.8
streamlit-webrtc>=0.47.0
